feat: CartRepository가 DAO 객체를 참조하도록 변경

+++ app/src/main/java/woowacourse/shopping/ui/cart/CartProductAdapter.kt
@@ -6,13 +6,13 @@ import woowacourse.shopping.model.Product
 
 class CartProductAdapter(
     items: List<Product>,
-    onClickDelete: (position: Int) -> Unit,
+    onClickDelete: (position: Long) -> Unit,
     private val dateFormatter: DateFormatter,
 ) : RecyclerView.Adapter<CartProductViewHolder>() {
     private val items: MutableList<Product> = items.toMutableList()
 
     private val onClickDelete = { position: Int ->
-        onClickDelete(position)
+        onClickDelete(position.toLong())
         removeItem(position)
     }
